Updated tests PMS persistency
Replaced kubeproxy with more efficient proxy

Issue-ID: NONRTRIC-486

+++ test/common/agent_api_functions.sh
@@ -270,9 +270,13 @@ start_policy_agent() {
 		fi
 
 		# Keep the initial worker node in case the pod need to be "restarted" - must be made to the same node due to a volume mounted on the host
-		__PA_WORKER_NODE=$(kubectl get pod -l "autotest=PA" -n $KUBE_NONRTRIC_NAMESPACE -o jsonpath='{.items[*].spec.nodeName}')
-		if [ -z "$__PA_WORKER_NODE" ]; then
-			echo -e $YELLOW" Cannot find worker node for pod for $POLICY_AGENT_APP_NAME, persistency may not work"$EYELLOW
+		if [ $retcode_i -eq 0 ]; then
+			__PA_WORKER_NODE=$(kubectl get pod -l "autotest=PA" -n $KUBE_NONRTRIC_NAMESPACE -o jsonpath='{.items[*].spec.nodeName}')
+			if [ -z "$__PA_WORKER_NODE" ]; then
+				echo -e $YELLOW" Cannot find worker node for pod for $POLICY_AGENT_APP_NAME, persistency may not work"$EYELLOW
+			fi
+		else
+			echo -e $YELLOW" Persistency may not work for app $POLICY_AGENT_APP_NAME in multi-worker node config when running it as a prestarted app"$EYELLOW
 		fi
 
 		echo " Retrieving host and ports for service..."
@@ -374,6 +378,13 @@ stop_policy_agent() {
 	echo -e $BOLD"Stopping $POLICY_AGENT_DISPLAY_NAME"$EBOLD
 
 	if [ $RUNMODE == "KUBE" ]; then
+
+		__check_prestarted_image "PA"
+		if [ $? -eq 0 ]; then
+			echo -e $YELLOW" Persistency may not work for app $POLICY_AGENT_APP_NAME in multi-worker node config when running it as a prestarted app"$EYELLOW
+			__kube_scale deployment $POLICY_AGENT_APP_NAME $KUBE_NONRTRIC_NAMESPACE 0
+			return 0
+		fi
 		__kube_scale_all_resources $KUBE_NONRTRIC_NAMESPACE autotest PA
 		echo "  Deleting the replica set - a new will be started when the app is started"
 		tmp=$(kubectl delete rs -n $KUBE_NONRTRIC_NAMESPACE -l "autotest=PA")
@@ -404,6 +415,14 @@ start_stopped_policy_agent() {
 
 	if [ $RUNMODE == "KUBE" ]; then
 
+		__check_prestarted_image "PA"
+		if [ $? -eq 0 ]; then
+			echo -e $YELLOW" Persistency may not work for app $POLICY_AGENT_APP_NAME in multi-worker node config when running it as a prestarted app"$EYELLOW
+			__kube_scale deployment $POLICY_AGENT_APP_NAME $KUBE_NONRTRIC_NAMESPACE 1
+			__check_service_start $POLICY_AGENT_APP_NAME $PA_PATH$POLICY_AGENT_ALIVE_URL
+			return 0
+		fi
+
 		# Tie the PMS to the same worker node it was initially started on
 		# A PVC of type hostPath is mounted to PMS, for persistent storage, so the PMS must always be on the node which mounted the volume
 		if [ -z "$__PA_WORKER_NODE" ]; then
@@ -419,7 +438,6 @@ start_stopped_policy_agent() {
 			fi
 			__kube_scale deployment $POLICY_AGENT_APP_NAME $KUBE_NONRTRIC_NAMESPACE 1
 		fi
-
 	else
 		docker start $POLICY_AGENT_APP_NAME &> ./tmp/.dockererr
 		if [ $? -ne 0 ]; then
@@ -2169,6 +2187,25 @@ api_get_configuration() {
 			return 1
 		fi
 	fi
+
+	__log_test_pass
+	return 0
+}
+
+##########################################
+####     Reset types and instances    ####
+##########################################
+
+# Admin reset to remove all policies and services
+# All types and instances etc are removed - types and instances in a1 sims need to be removed separately
+# NOTE - only works in kubernetes and the pod should not be running
+# args: -
+# (Function for test scripts)
+
+pms_kube_pvc_reset() {
+	__log_test_start $@
+
+	__kube_clean_pvc $POLICY_AGENT_APP_NAME nonrtric policymanagementservice-vardata-pvc /var/policy-management-service/database
 
 	__log_test_pass
 	return 0
